Handle absolute path when creating terminal
On WSL (Ubuntu 22.04, more specifically) the terminal would fail to launch with a "The terminal process failed to launch". This is happening because the workspace path is being returned as a relative path, but `vscode.window.createTerminal` expects an absolute `cwd` parameter.

We solved that by simply appending a `/` at the front of the path in case it doesn't have one yet.

Closes #33